Add check before rm -r -f on temp dir

+++ src/Scripts/run_native_nupipeline.sh
@@ -74,6 +74,8 @@ do
 
     nupipeline indir=$archive$ObsID steminputs=nu$ObsID outdir=$clean$ObsID"/pipeline_out/" | tee -a "$log_file"
 
-    echo removing $TMP_DIR
-    rm -r -f $TMP_DIR
+    if [[ $(dirname $TMP_DIR) -ef /tmp/ ]]
+        echo removing $TMP_DIR
+        rm -r -f $TMP_DIR
+    fi
 done
